+++ MaRTA/src/Serialize/serialize.h
@@ -3,7 +3,7 @@
 
 #include "../structs/job.h"
 
-t_job *desserealizeJob(int fd, uint32_t id);
+t_job *desserealizeJob(int fd, uint16_t id);
 void serializeMapToOrder(int fd, t_map *map);
 void serializeReduceToOrder(int fd, t_reduce *reduce);
